Setup project to run with express.js
Changes were made to the package.json files to rename to project-three.  index.js (first draft developed by Kinza) was re-added to the project and incorporated through a new connection to index.js and a modified file structure to mirror the typical express.js file structure.  Next, Alison will be modifying links between files so the CSS and JavaScript is properly connected to the HTML files.

Co-Authored-By: khaque1 <46754241+khaque1@users.noreply.github.com>